authanticate_gmail.py
import os
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
import settings


def get_auth_credentials():

    # Load the credentials from the file or initiate the OAuth flow
    if os.path.exists('token.json'):
        creds = Credentials.from_authorized_user_file('token.json', settings.SCOPES)
    else:
        flow = InstalledAppFlow.from_client_secrets_file(settings.CREDENTIALS_FILE, settings.SCOPES)
        creds = flow.run_local_server(port=0)
        # Save the credentials for future use
        with open('token.json', 'w') as token:
            token.write(creds.to_json())
    return creds
