requirements.txt

google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0

psycopg2==2.9.5
